Replace all i32 with i64 to reduce complexity OOPS AND DELETE SUM SUN
Lots of numbers absolutely don't need an i64 but tbh they absolutely didn't need an i32 either. Making the data types minimal could be a later exercise.

Also deletes the previous sum sun handling, to be replaced in the next commit. I didn't mean to do that. Not going to fix it though.